Update Data Wrangling
In this part of my analysis, I created a new column called AgeGroupAtMarriage_Man and  AgeGroupAtMarriage_Woman.
The idea was to categorize each woman's age at the time of marriage into defined age brackets, like '20-30', '31-40', and so on. 
Instead of working with exact ages, I wanted to group them so I could better analyze trends and patterns across age ranges.
This kind of grouping makes it easier to compare insights for example, do women who marry in their 20s experience shorter or longer marriages? 
Or is there a higher divorce rate in a certain age group? 
It also helps a lot when building visuals in Power BI, like bar charts or heat maps, where grouped data is much more intuitive and meaningful.
Ultimately, the goal was to make the data more interpretable both for myself and for any stakeholders who would view the report 
and to support further analysis, like understanding behavior across different age segments or building predictive models.